FIX: fix the duration property of SourceSimulator to return a float
When the duration is not provided to the SourceSimulator, it is computed
from the events. Instead of a duration in seconds (a float), it current
returns an integer.

This commit removes the cast to int to always return a float.

+++ mne/simulation/source.py
@@ -353,7 +353,7 @@ class SourceSimulator():
         """Duration of the simulation"""
         # If not, the precomputed maximum last sample is used
         if self._duration is None:
-            return int(np.max(self._last_samples) * self._tstep)
+            return np.max(self._last_samples) * self._tstep
         return self._duration
 
     @property
